feat: 서비스 계층 추상화로 트랜잭션 적용 로직과 비즈니스 로직 분리
