+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,0 +1,33 @@
+package com.techcourse.service;
+
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.domain.UserHistory;
+
+public class AppUserService implements UserService {
+
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
+
+    public AppUserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+    }
+
+    public User findById(final long id) {
+        return userDao.findById(id);
+    }
+
+
+    public void insert(final User user) {
+        userDao.insert(user);
+    }
+
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        final var user = userDao.findById(id);
+        user.changePassword(newPassword);
+        userDao.update(user);
+        userHistoryDao.log(new UserHistory(user, createBy));
+    }
+}
