+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -24,7 +24,7 @@ class UserServiceTest {
     private DataSource dataSource;
 
     @BeforeEach
-    void setUp() throws SQLException {
+    void setUp() {
         DatabasePopulatorUtils.execute(DataSourceConfig.getInstance());
         dataSource = DataSourceConfig.getInstance();
         this.jdbcTemplate = new JdbcTemplate(dataSource);
@@ -36,7 +36,8 @@ class UserServiceTest {
     @Test
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, dataSource);
+        AppUserService appUserService = new AppUserService(userDao, userHistoryDao);
+        final var userService = new TxUserService(appUserService,dataSource);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -51,7 +52,8 @@ class UserServiceTest {
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         MockUserHistoryDao userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, dataSource);
+        AppUserService appUserService = new AppUserService(userDao, userHistoryDao);
+        final var userService = new TxUserService(appUserService,dataSource);
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
